python_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
transformers==4.36.2
pydantic==2.5.3
sentencepiece==0.1.99
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu 